Improved velocity tracker.
Moved the velocity tracker to the Lua part where it's easier to implement and
integrate with the arpeggiator. In particular, this also lets us output the
true velocities from the arpeggiator, so that recorded loops now always play
back at the right level.

In arpeggiator mode, the gain control now acts as a dry/wet control which
mixes between minvel/maxvel preset parameters ("dry" values, gain = 1) and the
calculated minavg/maxavg envelope values ("wet" values, gain = 0). In loop
playback, it becomes a real gain control which can be used to change the
output velocities of the loop by -3 to 3 dB.

To make all this work as expected, default for the gain value is now
0.5 (center value), which means a mix halfway between dry and wet velocities
in the arpeggiator, and unit gain in loop playback. The velocity tracker
toggle is gone, you can now switch the tracker on and off by setting the gain
to 0 and 1, respectively. The gain is saved in the preset as before.